Fix amphora agent related issues
The amphora agent runs in a controlled environment, which currently
is a nova instance.

The meaning of this is that the log path is pre-known (and actually,
currently hard-coded[1]) and the single configuration file it reads will
always be /etc/octavia/amphora-agent.conf, as reflected in the
systemd file[2] used to spawn the amphora-agent.

An additional fix included with this patch is the user we use to start
the amphora-agent. Similarly to Ubuntu and CentOS, we need to use root.
Be minded that this python agent runs in a nova instance and not in
OpenStack controller or compute nodes.

[1] https://github.com/openstack/octavia/blob/039395f7ee2a84c0c166507feedccaefae44a9ca/octavia/cmd/agent.py#L79-L80
[2] https://github.com/openstack/octavia/blob/bf8aac5561647d1c2287ca1517ffc2da79edb281/elements/amphora-agent/init-scripts/systemd/amphora-agent.service

Related: rhbz#1450631

Change-Id: I40f2fe2177ea70567d62881484aac98c852b8f52