Update patch set 2
Patch Set 2:

Aye, both have a specific directory to log in, so they can write in it.
But amphora-agent tries to write directly to /var/log, and is not allowed to do so (process does not run as root).

So the RPM should say amphora-agent is allowed to create and write to the log file, something like a "%ghost" directive http://ftp.rpm.org/max-rpm/s1-rpm-inside-files-list-directives.html

Patch-set: 2