Update patch set 4
Patch Set 4:

> gunicorn 18.0 is coming from base EL7, I would look at relaxing
 > hard-dep on 19.0 upstream, on first sight it seems improper that it
 > uses internal gunicorn classes and not just WSGI interface.

The reasoning for 19.0 is https://review.openstack.org/#/c/386758 (also in the commit msg).
It's been part of the Octavia code starting stable/ocata, written by a core member in Octavia.
Why can't we simply bump the package version to have it working the same as it does in u/s?

Patch-set: 4
Label: Code-Review=0