Bump python2-barbicanclient and add werkzeug
Some of the requirements.txt values do not match the required minimum
versions for Octavia. This was recently discovered with a new gate using
lower-constraints [1].

python2-barbicanclient is one of those requirements that do not match
the required minimum versions [2].

Werkzeug was simply missing in requirements.txt. Because it was a
dependency of some Octavia requirement, it was being installed, hence
didn't raise a red flag [3].

[1] https://review.openstack.org/#/c/553134/
[2] https://review.openstack.org/#/c/553138/
[3] https://review.openstack.org/#/c/553141/

Change-Id: I5f0d7472ed8265fae680df2505f25dfdb26c2e8f