Activate policy.json
Use the admin_or_owner-policy.json to bypass load-balancer RBAC
role requirement. More about this here[1].

This will allow a similar behavior to what we know from other
projects such as Cinder and Nova.

The patch will move the policy file and rename in a similar
fashion to what is being done here[2].

[1] https://github.com/openstack/octavia/blob/master/etc/policy/README.rst
[2] https://github.com/openstack/octavia/blob/8e2f7512c2c6b06b6f284edb7cfc6f548ae1085b/devstack/plugin.sh#L305-L306

Change-Id: I13aad428d291f407d84bdedb03e25b3d5c5a1315