Remove golang-based httpd binary
After the Tempest code has been removed from Octavia [1], there is
no point in creating this binary anymore as part of the Octavia package.
We already have it as a subpackage of octavia-tempest-plugin [2], and
that is what Tempest is requiring [3].

Removing this subpackage, we also confine the golang dependency to
the Octavia Tempest plugin package.

[1] - https://review.opendev.org/659516
[2] - https://github.com/rdo-packages/octavia-tempest-plugin-distgit/blob/rpm-master/python-octavia-tests-tempest.spec#L151-L153
[3] - https://github.com/rdo-packages/tempest-distgit/blob/rpm-master/openstack-tempest.spec#L147

Change-Id: Iff60c26439b6dfc5c88636ed9097935e4e918307